backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pgvector
sentence-transformers
PyMuPDF
pymupdf4llm
pdfplumber
python-multipart
python-dotenv
pydantic-settings
tiktoken
cryptography
openai
anthropic
